backend/requirements.txt
djangorestframework
django-import-export
django
django-extensions
autopep8
psycopg2-binary>=2.8
django-environ
graphene-django